complete collision check and abstract class for enemies

+++ my-game/Sprites.h
@@ -10,16 +10,16 @@ class CSprite
 {
 	string id;				// Sprite ID in the sprite database
 
-	int left;
-	int top;
-	int right;
-	int bottom;
+	float left;
+	float top;
+	float right;
+	float bottom;
 
 	LPDIRECT3DTEXTURE9 texture;
 public:
-	int width;
-	int height;
-	CSprite(string id, int left, int top, int right, int bottom, int width, int height, LPDIRECT3DTEXTURE9 tex);
+	float width;
+	float height;
+	CSprite(string id, float left, float top, float right, float bottom, float width, float height, LPDIRECT3DTEXTURE9 tex);
 
 	void Draw(float x, float y, int alpha = 255);
 	void DrawWithScale(float x, float y, Vector scale = Vector(1.0f, 1.0f), int alpha = 255);
@@ -37,7 +37,7 @@ class CSprites
 	unordered_map<string, LPSPRITE> sprites;
 
 public:
-	void Add(string id, int left, int top, int right, int bottom, LPDIRECT3DTEXTURE9 tex);
+	void Add(string id, float left, float top, float right, float bottom, LPDIRECT3DTEXTURE9 tex);
 	LPSPRITE Get(string id);
 	void CSprites::Clear();
